+++ client/src/pages/AccountPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AccountPage = () => {
+  return (
+    <div>AccountPage</div>
+  )
+}
+
+export default AccountPage
